requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
argon==0.2.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
assem==0.3.1
asyncpg==0.30.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg==3.2.11
psycopg-binary==3.2.11
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
